requirements.txt
streamlit==1.28.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
psycopg2-binary==2.9.7
plotly==5.15.0
seaborn==0.12.2
matplotlib==3.7.2
mlflow==2.6.0
scipy==1.11.1
sqlalchemy==2.0.20
python-dotenv==1.0.0
joblib==1.3.2